Hems | upload is done

+++ server/src/middlewares/uploadMiddlware2.js
@@ -4,16 +4,19 @@ import fs from "fs";
 
 const uploadPhoto = async (req, res, next) => {
   try {
+    if(!req.file){
+      return res.status(400).json({message: "Upload the file"});
+    }
     const result = await uploadFunction(req.file.path);
     fs.unlinkSync(req.file.path);
     const createdPhoto = await photoModel.create({ photoId: result.public_id, url: result.secure_url });
     res.uploadedPhoto = createdPhoto;
     next();
   } catch (err) {
-    // console.log(err.error.Error)
-    if (err.error.errno === -3008) {
-      return res.status(400).json({message: `Yo bro, You are offline`});
-    }
+    console.log(err);
+    // if (err.error.errno === -3008) {
+    //   return res.status(400).json({message: `Yo bro, You are offline`});
+    // }
     res.status(400).json({ message: "Internal Server Error" });
   }
   // res.json({file: req.file, result,message: "uploaded successfully"})
